Add support for documentHighlight in vim8
